+++ source/include/core/common.h
@@ -1,0 +1,9 @@
+#ifndef SARDA_COMMON_H
+#define SARDA_COMMON_H
+
+#include <iostream> // for core/modules/logger.h
+#include <sstream> // for core/modules/logger.h
+#include <string>
+#include <time.h> // for core/utils/time.h
+
+#endif