+++ source/include/core/config.h
@@ -1,0 +1,9 @@
+#ifndef SARDA_CONFIG_H
+#define SARDA_CONFIG_H
+
+namespace sarda
+{
+	
+}
+
+#endif