+++ source/include/core/core.h
@@ -1,0 +1,7 @@
+#ifndef SARDA_CORE_H
+#define SARDA_CORE_H
+
+#include <include/core/proto-core.h>
+#include <include/core/modules/logger.h>
+
+#endif