+++ source/include/core/proto-core.h
@@ -1,0 +1,11 @@
+#ifndef SARDA_PROTO_CORE_H
+#define SARDA_PROTO_CORE_H
+
+#include <include/core/common.h>
+
+namespace sarda
+{
+	
+}
+
+#endif