+++ source/include/sarda.h
@@ -1,0 +1,11 @@
+#ifndef SARDA_SARDA_H
+#define SARDA_SARDA_H
+
+#include <include/core/core.h>
+
+namespace sarda
+{
+	
+}
+
+#endif