+++ NEWS.md
@@ -1,2 +1,3 @@
 ## SARDA 0.0.1 ##
++ Implemented a simple logger based on color-console library
 + Initial commit